Moved metadata tests into metadata module.
