+++ src/metadata.rs
@@ -1,6 +1,6 @@
 use crate::errors::*;
 use crate::gdal_major_object::MajorObject;
-use crate::utils::{_last_cpl_err, _last_null_pointer_err, _string};
+use crate::utils::{_last_cpl_err, _last_null_pointer_err, _string, _string_array};
 use gdal_sys::{self, CPLErr};
 use std::ffi::CString;
 
@@ -128,13 +128,106 @@ pub trait Metadata: MajorObject {
         Ok(())
     }
 
+    /// For Datasets this sets the dataset name; normally
+    /// application code should not set the "description" for
+    /// GDALDatasets. For RasterBands it is actually a description
+    /// (if supported) or "".
     fn set_description(&mut self, description: &str) -> Result<()> {
-        // For Datasets this sets the dataset name; normally
-        // application code should not set the "description" for
-        // GDALDatasets. For RasterBands it is actually a description
-        // (if supported) or "".
+
         let c_description = CString::new(description.to_owned())?;
         unsafe { gdal_sys::GDALSetDescription(self.gdal_object_ptr(), c_description.as_ptr()) };
         Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::*;
+
+    #[test]
+    fn test_get_dataset_driver() {
+        let dataset = Dataset::open(fixture!("tinymarble.png")).unwrap();
+        let driver = dataset.driver();
+        assert_eq!(driver.short_name(), "PNG");
+        assert_eq!(driver.long_name(), "Portable Network Graphics");
+    }
+
+    #[test]
+    fn test_get_description() {
+        let driver = Driver::get_by_name("mem").unwrap();
+        assert_eq!(driver.description().unwrap(), "MEM".to_string());
+    }
+
+    #[test]
+    fn test_get_metadata_domains() {
+        let dataset = Dataset::open(fixture!("tinymarble.png")).unwrap();
+        let mut domains = dataset.metadata_domains();
+        if domains[0].is_empty() {
+            domains.remove(0);
+        }
+
+        assert_eq!(
+            domains,
+            vec!(
+                "IMAGE_STRUCTURE",
+                "xml:XMP",
+                "DERIVED_SUBDATASETS",
+                "COLOR_PROFILE"
+            )
+        );
+    }
+
+    #[test]
+    fn test_get_metadata_domain() {
+        let dataset = Dataset::open(fixture!("tinymarble.png")).unwrap();
+        let domain = "None";
+        let meta = dataset.metadata_domain(domain);
+        assert_eq!(meta, None);
+
+        let domain = "IMAGE_STRUCTURE";
+        let meta = dataset.metadata_domain(domain);
+        assert_eq!(meta, Some(vec!(String::from("INTERLEAVE=PIXEL"))));
+    }
+
+    #[test]
+    fn test_get_metadata_item() {
+        let dataset = Dataset::open(fixture!("tinymarble.png")).unwrap();
+        let key = "None";
+        let domain = "None";
+        let meta = dataset.metadata_item(key, domain);
+        assert_eq!(meta, None);
+
+        let key = "INTERLEAVE";
+        let domain = "IMAGE_STRUCTURE";
+        let meta = dataset.metadata_item(key, domain);
+        assert_eq!(meta, Some(String::from("PIXEL")));
+    }
+
+    #[test]
+    fn test_set_metadata_item() {
+        let driver = Driver::get_by_name("MEM").unwrap();
+        let mut dataset = driver.create("", 1, 1, 1).unwrap();
+
+        let key = "Test_Key";
+        let domain = "Test_Domain";
+        let value = "Test_Value";
+        let result = dataset.set_metadata_item(key, value, domain);
+        assert!(result.is_ok());
+
+        let result = dataset.metadata_item(key, domain);
+        assert_eq!(Some(value.to_owned()), result);
+    }
+
+    #[test]
+    fn test_set_description() {
+        let driver = Driver::get_by_name("MEM").unwrap();
+        let dataset = driver.create("", 1, 1, 1).unwrap();
+        let mut band = dataset.rasterband(1).unwrap();
+
+        let description = "A merry and cheerful band description";
+        assert_eq!(band.description().unwrap(), "");
+
+        band.set_description(description).unwrap();
+        assert_eq!(band.description().unwrap(), description);
     }
 }
